Added conditionals to select some clauses, also added some comments.
Added conditionals to handle the open-source vs. closed-source clauses,
the retainer clause, and the end date.

Added some comments in the clientsinfo.tex file to assist in how to use.

There are MANY improvments to be had.
TODO: Find a better way to handle conditionals.  At the moment, mutually
exclusive is really the only option.  A proper way would be great.
TODO: Have the open-source clause reviewed, as it was written quickly.